refactor : 로그인 / 세션 불일치 문제 브랜치 분리 및 아키텍쳐 구조 수정
 1. 로그인 구현과 세션 불일치 문제를 해결하기 위한 세션 스토리지 추가 이슈 분리
 2. 팀원과 아키텍쳐 구조 통일
 3. HttpStatus 상수 이름 변경
 4. 리뷰반영(my-info - > my-infos : URL은 복수여야 한다.)
