+++ src/main/java/com/flab/shoeauction/common/configs/WebConfig.java
@@ -1,5 +1,9 @@
 package com.flab.shoeauction.common.configs;
 
+/**
+ *
+ */
+
 import com.flab.shoeauction.common.interceptor.LoginCheckInterceptor;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -8,12 +12,17 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+/**
+ * WebMvcConfigurer : 스프링 부트가 기본으로 설정한 MVC설정에 추가적으로 기능을 커스터마이징 할 수 있다.
+ */
 @Configuration
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
+
     private final LoginCheckInterceptor loginCheckInterceptor;
     private final LoginUserArgumentResolver loginUserArgumentResolver;
+
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(loginCheckInterceptor);
