+++ src/main/java/com/flab/shoeauction/common/utils/user/UserConstants.java
@@ -1,4 +1,4 @@
-package com.flab.shoeauction.user.utils;
+package com.flab.shoeauction.common.utils.user;
 
 public class UserConstants {
     public static final String USER_ID = "email";
