+++ src/test/java/com/flab/shoeauction/service/SignUpServiceTest.java
@@ -1,4 +1,4 @@
-package com.flab.shoeauction.user.service;
+package com.flab.shoeauction.service;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.any;
@@ -6,11 +6,11 @@ import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.flab.shoeauction.user.service.encrytion.EncryptionService;
-import com.flab.shoeauction.user.dto.UserDto;
-import com.flab.shoeauction.user.exception.EmailDuplicateException;
-import com.flab.shoeauction.user.exception.NicknameDuplicateException;
-import com.flab.shoeauction.user.repository.UserRepository;
+import com.flab.shoeauction.service.encrytion.EncryptionService;
+import com.flab.shoeauction.web.dto.UserDto;
+import com.flab.shoeauction.exception.user.EmailDuplicateException;
+import com.flab.shoeauction.exception.user.NicknameDuplicateException;
+import com.flab.shoeauction.domain.user.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
